file structure refactoring in keyboard and caret
