+++ package.json
@@ -35,10 +35,9 @@
     "fs-extra": "^11.2.0",
     "ts-node": "^10.9.2",
     "typescript": "^5.4.5",
-    "typescript-eslint": "^7.9.0"
+    "typescript-eslint": "^7.9.0",
+    "vitest": "1.4.0"
   },
   "packageManager": "yarn@4.4.0",
-  "dependencies": {
-    "vitest": "1.4.0"
-  }
+  "dependencies": {}
 }
