+++ packages/caret/src/checkContenteditableSliceForEmptiness/index.ts
@@ -1,0 +1,3 @@
+import { checkContenteditableSliceForEmptiness } from './checkContenteditableSliceForEmptiness';
+
+export { checkContenteditableSliceForEmptiness };
