+++ packages/caret/src/focus/index.ts
@@ -1,0 +1,3 @@
+import { focus } from './focus';
+
+export { focus };
