+++ packages/caret/src/getCaretNodeAndOffset/getCaretNodeAndOffset.ts
