+++ packages/caret/src/getCaretNodeAndOffset/index.ts
@@ -1,0 +1,3 @@
+import { getCaretNodeAndOffset } from './getCaretNodeAndOffset';
+
+export { getCaretNodeAndOffset };
