+++ packages/caret/src/getContenteditableSlice/index.ts
@@ -1,0 +1,3 @@
+import { getContenteditableSlice } from './getContenteditableSlice';
+
+export { getContenteditableSlice };
