+++ packages/caret/src/getRange/getRange.ts
