+++ packages/caret/src/getRange/index.ts
@@ -1,0 +1,3 @@
+import { getRange } from './getRange';
+
+export { getRange };
