+++ packages/caret/src/isCaretAtEndOfInput/index.ts
@@ -1,0 +1,3 @@
+import { isCaretAtEndOfInput } from './isCaretAtEndOfInput';
+
+export { isCaretAtEndOfInput };
