+++ packages/caret/src/isCaretAtStartOfInput/index.ts
@@ -1,0 +1,3 @@
+import { isCaretAtStartOfInput } from './isCaretAtStartOfInput';
+
+export { isCaretAtStartOfInput };
