+++ packages/keyboard/src/getKeyboardKeyForCode/index.ts
@@ -1,0 +1,3 @@
+import { getKeyboardKeyForCode } from './getKeyboardKeyForCode';
+
+export { getKeyboardKeyForCode };
